+++ test/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './InstantSearchHooksTestWrapper';
 export * from './createInstantSearchTestWrapper';
 export * from './runAllMicroTasks';
 export * from './wait';
